# Different bitmap sizes support: speedup 'And', 'AndNot', fix 'AndNot', 'Or' and 'Xor'